backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-dotenv>=1.0.0

# Database dependencies
mysql-connector-python>=8.2.0
sqlalchemy>=2.0.23

# Authentication and security
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Data processing and analysis
pandas>=2.1.0
numpy>=1.25.0

# OpenAI API for chat features
openai>=1.0.0

# Machine Learning Dependencies
scikit-learn>=1.3.2
joblib>=1.3.2

# HTTP client for API calls
requests>=2.31.0